+++ src/infi/clickhouse_orm/models.py
@@ -1,7 +1,8 @@
 from __future__ import unicode_literals
+import sys
 from logging import getLogger
 
-from six import with_metaclass
+from six import with_metaclass, reraise
 import pytz
 
 from .fields import Field, StringField
@@ -124,8 +125,13 @@ class Model(with_metaclass(ModelBase)):
         '''
         field = self.get_field(name)
         if field:
-            value = field.to_python(value, pytz.utc)
-            field.validate(value)
+            try:
+                value = field.to_python(value, pytz.utc)
+                field.validate(value)
+            except ValueError:
+                tp, v, tb = sys.exc_info()
+                new_msg = "{} (field '{}')".format(v, name)
+                reraise(tp, tp(new_msg), tb)
         super(Model, self).__setattr__(name, value)
 
     def set_database(self, db):
@@ -274,3 +280,9 @@ class MergeModel(Model):
         res = super(MergeModel, self).set_database(db)
         self.engine.set_db_name(db.db_name)
         return res
+
+    @classmethod
+    def create_table_sql(cls, db_name):
+        assert isinstance(cls.engine, Merge), "engine must be engines.Merge instance"
+        cls.engine.set_db_name(db_name)
+        return super(MergeModel, cls).create_table_sql(db_name)